python/bots/requirements.txt
pyTelegramBotAPI>=4.0
discord.py>=1.7